rental + work evidence

+++ src/main/java/machineRental/MR/repository/MachineTypeRepository.java
@@ -1,11 +1,11 @@
 package machineRental.MR.repository;
 
-import machineRental.MR.model.MachineType;
+import machineRental.MR.machineType.model.MachineType;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MachineTypeRepository extends JpaRepository<MachineType, Integer> {
+public interface MachineTypeRepository extends JpaRepository<MachineType, Long> {
 
     boolean existsByMachineType(String machineType);
